@@ -4,7 +4,7 @@
 	"type": "library",
 	"license": "LGPL-3.0",
 	"require": {
-		"php": "^7.4 || ^8.0",
+		"php": "^8.1",
 		"php-64bit": "*"
 	},
 	"require-dev": {
@@ -17,7 +17,10 @@
 	"autoload": {
 		"psr-4": {
 			"pocketmine\\utils\\": "src/"
-		}
+		},
+		"files": [
+			"src/bootstrap.php"
+		]
 	},
 	"autoload-dev": {
 		"psr-4": {
